move auto renewal method to helper file
