+++ vault/resource_pki_secret_backend_sign.go
@@ -34,8 +34,9 @@ func pkiSecretBackendSignResource() *schema.Resource {
 			},
 		},
 		SchemaVersion: 1,
-		CustomizeDiff: pkiCertAutoRenewCustomizeDiff,
-
+		CustomizeDiff: func(_ context.Context, d *schema.ResourceDiff, meta interface{}) error {
+			return pkiCertPlanAutoRenewal(d)
+		},
 		Schema: map[string]*schema.Schema{
 			consts.FieldBackend: {
 				Type:        schema.TypeString,
